% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting_viz.R
\name{sim_bets}
\alias{sim_bets}
\title{Simulate Betting Outcomes}
\usage{
sim_bets(n_bets, n_sim, edge, mean_stake = 100, same_odds = NULL)
}
\arguments{
\item{n_bets}{An integer specifying the number of bets to simulate.}

\item{n_sim}{An integer specifying the number of simulations to perform.}

\item{edge}{A numeric value representing the edge you have (e.g., 1.10 for 10 percent edge).}

\item{mean_stake}{A numeric value for the mean stake amount per bet, default is 100.}

\item{same_odds}{A numeric value or vector to set the same odds for all bets, default is NULL.}
}
\value{
A list with:
\item{profit_plot}{A ggplot object of cumulative profit across bets.}
\item{odds_plot}{A ggplot object of the odds distribution.}
}
\description{
This function simulates the cumulative profit over a specified number of bets and simulations,
allowing for the calculation of expected profit based on given odds and stake.
}
\examples{
sim_bets(n_bets = 1000, n_sim = 20, edge = 1.05)
}
